refactor(most utils): regrouped most.js related utilities
